requirements.ml-onnx.txt
onnxruntime>=1.19
Pillow
numpy
huggingface-hub
